Update 2_functional-tests.js
This additional code was added to give me clarification on how to update the code using copilot.  I hope this clears up any confusion. Was also wondering if I could use copilot to finish and complete the rest of the code for the Stock Price Checker Project.  And do I need a API or IP Address to add to my code.